+++ etils/ecolab/__init__.py
@@ -20,6 +20,7 @@ from etils.ecolab.array_as_img import auto_plot_array
 from etils.ecolab.auto_display_utils import auto_display
 from etils.ecolab.auto_display_utils import disp
 from etils.ecolab.colab_utils import collapse
+from etils.ecolab.colab_utils import get_permalink
 from etils.ecolab.colab_utils import interruptible
 from etils.ecolab.colab_utils import json
 from etils.ecolab.highlight_util import highlight_html